добавлена валидация количества комнат и мест
