+++ css/style.css
@@ -1014,6 +1014,10 @@ fieldset {
   opacity: 0.6;
 }
 
+.ad-form__element--error {
+  color: #ff0000;
+}
+
 .visually-hidden {
   position: absolute;
