+++ js/main.js
@@ -1,5 +1,5 @@
 import './data.js';
 import './util.js';
 import './generate.js';
-import './formToggle.js';
-import './formValidating.js';
+import './form-toggle.js';
+import './form-validating.js';
